+++ CHANGELOG.md
@@ -9,6 +9,8 @@ experienced users of SimFin.
 
 -   Added bulk-data load-functions for banks and insurance companies.
 -   Changed all specialized bulk-data load-functions to use functools.partial.
+-   Added 'full' variants for datasets: income, balance, cashflow.
+    The 'full' variant has more columns than the 'free' variant.
 
 
 ## Version 0.1.0 (2019-09-26)